refactor: move model to api package, conslidate tool definition structures and introduce jsonx package for dynamic JSON conversion
