+++ pkg/jsonx/dynamic_test.go
@@ -1,0 +1,50 @@
+package jsonx
+
+import (
+	"testing"
+
+	"github.com/stretchr/testify/assert"
+)
+
+func TestToDynamicJSON(t *testing.T) {
+	tests := []struct {
+		name    string
+		input   interface{}
+		want    map[string]interface{}
+		wantErr bool
+	}{
+		{
+			name: "simple struct",
+			input: struct {
+				Name string `json:"name"`
+				Age  int    `json:"age"`
+			}{
+				Name: "test",
+				Age:  30,
+			},
+			want: map[string]interface{}{
+				"name": "test",
+				"age":  float64(30),
+			},
+			wantErr: false,
+		},
+		{
+			name:    "invalid input",
+			input:   make(chan int),
+			want:    nil,
+			wantErr: true,
+		},
+	}
+
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			got, err := ToDynamicJSON(tt.input)
+			if tt.wantErr {
+				assert.Error(t, err)
+				return
+			}
+			assert.NoError(t, err)
+			assert.Equal(t, tt.want, got)
+		})
+	}
+}
