+++ provider/openai/model.go
@@ -4,32 +4,32 @@ import (
 	"sync"
 
 	"github.com/alphadose/haxmap"
-	"github.com/casualjim/bubo"
+	"github.com/casualjim/bubo/api"
 	"github.com/casualjim/bubo/provider"
 	"github.com/openai/openai-go"
 	"github.com/openai/openai-go/option"
 )
 
-var modelRegistry = haxmap.New[string, bubo.Model]()
+var modelRegistry = haxmap.New[string, api.Model]()
 
-func GPT4oMini(opts ...option.RequestOption) bubo.Model {
+func GPT4oMini(opts ...option.RequestOption) api.Model {
 	return Model(openai.ChatModelGPT4oMini, opts...)
 }
 
-func GPT4o(opts ...option.RequestOption) bubo.Model {
+func GPT4o(opts ...option.RequestOption) api.Model {
 	return Model(openai.ChatModelChatgpt4oLatest, opts...)
 }
 
-func O1Mini(opts ...option.RequestOption) bubo.Model {
+func O1Mini(opts ...option.RequestOption) api.Model {
 	return Model(openai.ChatModelO1Mini, opts...)
 }
 
-func O1(opts ...option.RequestOption) bubo.Model {
+func O1(opts ...option.RequestOption) api.Model {
 	return Model(openai.ChatModelO1, opts...)
 }
 
-func Model(name string, opts ...option.RequestOption) bubo.Model {
-	m, _ := modelRegistry.GetOrCompute(name, func() bubo.Model {
+func Model(name string, opts ...option.RequestOption) api.Model {
+	m, _ := modelRegistry.GetOrCompute(name, func() api.Model {
 		return &model{
 			name: name,
 			opts: opts,
@@ -38,7 +38,7 @@ func Model(name string, opts ...option.RequestOption) bubo.Model {
 	return m
 }
 
-var _ bubo.Model = (*model)(nil)
+var _ api.Model = (*model)(nil)
 
 type model struct {
 	name string
